Let's get dangerous!
This file works miracles.  The primary problem with the mozilla port is the JavaScript Engine.  This change makes the JS engine work on amd64 provided the makefile is hacked for amd64 instead of i386.

make fake isn't 100% yet and make package still fails but if you manually run ./mozilla as root in the wrksrc/fake or possible the fake root (there actually is a fake directory created by the port unrelated to our fake) it actually works!  I was able to load mozilla.org with it.

Note all i had enabled for the build was DEBUG and the browser.  Enabling other things (like ldap) causes it to fail still.