Update to 1.2.18.
Fix long standing problem where libpng-config was created in /usr/local/bin instead of the fake directory.  As the path is still set properly to /usr/local for the library in libpng-conf but it is installed in ${DESTDIR}${PREFIX}, everything should work according to plan.

The cairo  port is acting up with png, so hopefully we've fixed some of the problems on that front.