Overhaul on bsd.apache.mk.
Remove most of the APACHE 1.3 stuff since we never intend to support that outdated version.  Many people have issues with the newer apache license, but they often switch to lighttpd, bozo or some other webserver.

The other reasoning is that MidnightBSD is primarily a desktop operating system and so we need to support development which is going to focus on the future i.e. Apache 2.x.

Worst case this could always be merged back in if someone actually has the energy to port apache 1.3.x.

We should sweep for apache 1.3+ usages for versions and correct them.