--infodir and --mandir are now set for gnu configure scripts, provided the script supports the arguments.
new FAKE_OPTS option: prefixhack.  This implements the common idium of
overiding prefix when make install is called in the wrksrc.  It makes sure
to override infodir and mandir, as this are now explicitly set and no longer
$prefix relative.

plists generated by makeplist (gen-plist) now use @dirrm instead of
@dirrmtry.

Continuing changes to describe-yaml, as this is still a moving target.