Update libpurple to 2.2.1 from pidgin.
This is not being used as a MASTER port for pidgin anymore.  The PIDGIN_SLAVE concept was seriously broken with our current ports system.  This will be fixed at a later time.  We're disabling it with gaim or pidgin installed and leaving the port in case we want it for another program.

We should revisit the slave problem later.