Broke the common commands into bsd.commands.mk.
Documentation fixes.  The docs in the begining of bsd.mport.mk now reflect
reality.

Added SKIP_FAKE_CHECK.  A list of files that aren't checked to see if they
contain the fake destdir.  Perl now adds the packlist to this automatically.

Replaced TARGETDIR with PREFIX in many places.  This is a continuation of
work to restrict the use of DESTDIR to install-package.

USE_XORG now works properly with pre/post.port.mk.