Update to 2.4.2.
Add missing dependancy on libxml2.Explicitly disable several features if they are not enabled such as avahi, meanwhile, etc.

From their changelog:

libpurple
In MySpaceIM, messages from spambots are discarded (Justin Williams)
Strip mIRC formatting codes from quit and part messages.
IRC now displays ban lists in-channel for joined channels.
Fixed a bug where the list of loaded plugins would get removed when switching between different operating systems.
Fix reception of IRC PART without a part message on Undernet (fixes a problem with litter in the channel user list).
IRC no longer crashes on /list on servers which erroneously omit RPL_LISTSTART.
Update the NetworkManager support to use D-Bus directly, instead of libnm-glib. Hopefully it's stable now. It will now compile by default if you have D-Bus support and NetworkManager.h. (Elliott Sales de Andrade)
MSN buddy list synchronization is now more forgiving, only asking about buddies who have disappeared completely from the server list and not those that have simply moved groups.
IRC will now try to append 1-9 to your nick if it is in use, instead of substituting the last character with 1-9 where possible.
Bonjour buddies will be saved persistently if they're moved out of the "Bonjour" group. (Eion Robb)
Pidgin
The typing notification in the conversation history can be disabled or customized (font, color etc.) in .gtkrc-2.0.
Added a plugin (not installed by default) which adds a Send button back to the conversation window. People without physical keyboards have a hard time with the lack of the button.
Clicking on the buddyicon in the conversation window toggles the size of the icon between small and large.
The settings of a chat (e.g. Handle in an XMPP chat, or Exchange in an AIM chat) can be edited from its context menu in the buddy list.
Add a "Present conversation window" preference to the Message Notification plugin; the "Raise conversation window" option does not unminimize windows or draw attention to them when they are on other workspaces--the "Present" option should.
Add a preference to set Escape as the keyboard shortcut for closing the conversation window.
Add an option in the context menu to disable smileys in the selected text in the conversation history/log viewer. This should help people who regularly paste code in conversations.
Add a preference to choose the minimum size of the text input area in lines.
Moved the "Local alias" field in the Modify Account dialog to be below the "User Options" heading on the "Basic" tab.
Number of room occupants is now shown in chat tooltips where possible
General
The configure script now dies on more absent dependencies. The --disable-xxx arguments to configure can be used to bypass unneeded dependencies. This will also cause the configure script to die if an --enable-xxx option is used and the dependencies it requires are missing.
The Evolution integration plugin must now be explicitly enabled. Use the --enable-gevolution argument to configure to enable it.
The Contact Availability Prediction plugin must now be explicitly enabled. Use the --enable-cap argument to configure to enable it.