Move the adobe acrobat german ports to print.  This will keep them together.  Some time ago, ctriv and I talked about not using seperate directories for language support.  It's awkward as you have to look there first to see if the app is in your native language and then elsewhere if you fail to find something.
Hopefully mport's gui will sort by categories so we can just tag them with the language like german, etc.  (maybe that should be i18n too)

For now, follow this convention until we decide on something better.

On the file system, the lang is the suffix, but it's the PKGNAMEPREFIX in the port file.  There are situations we can't use a suffix nore often than a prefix.  However, on the file system it sucks not to see all the acrobat ports together for instance.  Perhaps we should make a special prefix or suffix on packages for langauges?  I'll let ctriv think about that.