Magus claims this is installing something outside the fakedir (finaldir) but I can't figure out what it is.
There isn't an install target with the included Makefile and the datadir and docsdir stuff looks good.  I've used @ to silence the output after i verified it's accuracy.

There could be something named similar or something in the magus tarball that is being picked up?  I need to look at how magus does the check to narrow it down.