After thinking about this for a minute, the if statements were reversered.
In the case where sqlite is defined as yes or 3, we want the dependancy added for 0.2.x and lower, but not current as it contains sqlite3.

In the case the port has an INVALID version, we want to error out as this is just wrong.  The previous logic would only complain in the event it was set to an invalid number on an old version of midnightbsd.  Since developers run current most often, this is silly.

One could argue we don't need the version check at all, but if we ever need to add more versions to mports it's nice to have the logic ready. (sqlite 2 or even 4 in the future?)