gnustep examples 1.4.0
This is rather hackish as CurrencyConverter example is not compiling and I read they were trying to work on modernizing some of the samples.  It could either be due to us not running development gnustep or that the sample has not been updated yet.

Adding "." back to the list of directories will bring in everything missing.