Move definition of MAKE_CMD.
We now define it in bsd.mport.mk rather than commands.mk.
it must be loaded before fake targets so we moved it up the file a bit from the other make related definitions.

This more closely aligns with FreeBSD ports and works with the emacs-devel port trying to load BSDMAKE.