expat 2.7.0.  fixes cve-2024-8176.
also update a few extensions to improve formatting and
fix bugs with some.

Added an extension form ravenports that they use on midnightbsd
builds for fixing autotools files.  We have some logic for some of this
also, but this seems cleaner for some cases.  I'm not using it yet
and we likely need to fix the other behavior. (it may go away if we
improve the native logic)